% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmfgpu4R.R
\name{chooseGpu}
\alias{chooseGpu}
\title{Selects the specified device as primary computation device. All further invocations to nmfgpu will use the specified
CUDA device.}
\usage{
chooseGpu(device.index)
}
\arguments{
\item{device.index}{Index of the CUDA device, which should be used for computation.}
}
\description{
Selects the specified device as primary computation device. All further invocations to nmfgpu will use the specified
CUDA device.
}

