% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmfgpu4R.R
\name{print.DeviceMemoryInfo}
\alias{print.DeviceMemoryInfo}
\title{Prints the information of a 'DeviceMemoryInfo' object.}
\usage{
\method{print}{DeviceMemoryInfo}(x, ...)
}
\arguments{
\item{x}{Object of class 'DeviceMemoryInfo'}

\item{...}{Other arguments}
}
\description{
Prints the information of a 'DeviceMemoryInfo' object.
}

