% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nmfgpu4R.R
\name{cudaMemoryInfo}
\alias{cudaMemoryInfo}
\title{Test}
\usage{
cudaMemoryInfo(device.index = NA)
}
\arguments{
\item{device.index}{If specified the memory info retrieval is restricted to the passed device indices. By default no restriction is active and
therefore memory information about all available CUDA devices are retrieved.}
}
\description{
Test
}

