% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmfgpu4R.R
\docType{package}
\name{nmfgpu4R}
\alias{nmfgpu4R}
\alias{nmfgpu4R-package}
\title{R binding for computing non-negative matrix factorizations using CUDA}
\description{
R binding for the libary \emph{nmfgpu} which can be used to compute Non-negative Matrix Factorizations (NMF) using CUDA hardware
acceleration.
}
\details{
The main function to use is \code{\link{nmfgpu}} which can be configured using various arguments.
In addition to it a few helper functions are provided, but they aren't necessary for using \code{\link{nmfgpu}}.
}

